+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Bumps Instabug iOS SDK to v10.11.8
 * Bumps Instabug Android SDK to v10.13.0
+* Adds Instabug Experiments APIs
 * Adds defensive type checking in Instabug logging APIs
 * Excludes unnecessary files from the published npm package
 * Fixes an issue with GraphQL requests not being grouped correctly